Added a paragraph for saving new posts as drafts.
In the Django Forms tutorial chapter, the post.published_date member is set when we
create the post.

This commit adds a short paragraph on removing that line which will allow us to create
drafts that we later can publish with the publish button, making it more understandable
to why we need to be able to list and publish drafts in the first hand.